+++ astropy/table/operations.py
@@ -2,6 +2,7 @@
 High-level table operations:
 
 - join()
+- setdiff()
 - hstack()
 - vstack()
 """
@@ -22,7 +23,7 @@ from .column import Column
 from . import _np_utils
 from .np_utils import fix_column_name, TableMergeError
 
-__all__ = ['join', 'hstack', 'vstack', 'unique']
+__all__ = ['join', 'setdiff', 'hstack', 'vstack', 'unique']
 
 
 def _merge_table_meta(out, tables, metadata_conflicts='warn'):
@@ -123,6 +124,100 @@ def join(left, right, keys=None, join_type='inner',
     _merge_table_meta(out, [left, right], metadata_conflicts=metadata_conflicts)
 
     return out
+
+
+def setdiff(table1, table2, keys=None):
+    """
+    Take a set difference of table rows.
+
+    The row set difference will contain all rows in ``table1`` that are not
+    present in ``table2``. If the keys parameter is not defined, all columns in
+    ``table1`` will be included in the output table.
+
+    Parameters
+    ----------
+    table1 : `~astropy.table.Table`
+        ``table1`` is on the left side of the set difference.
+    table2 : `~astropy.table.Table`
+        ``table2`` is on the right side of the set difference.
+    keys : str or list of str
+        Name(s) of column(s) used to match rows of left and right tables.
+        Default is to use all columns in ``table1``.
+
+    Returns
+    -------
+    diff_table : `~astropy.table.Table`
+        New table containing the set difference between tables. If the set
+        difference is none, an empty table will be returned.
+
+    Examples
+    --------
+    To get a set difference between two tables::
+
+      >>> from astropy.table import setdiff, Table
+      >>> t1 = Table({'a': [1, 4, 9], 'b': ['c', 'd', 'f']}, names=('a', 'b'))
+      >>> t2 = Table({'a': [1, 5, 9], 'b': ['c', 'b', 'f']}, names=('a', 'b'))
+      >>> print(t1)
+       a   b
+      --- ---
+        1   c
+        4   d
+        9   f
+      >>> print(t2)
+       a   b
+      --- ---
+        1   c
+        5   b
+        9   f
+      >>> print(setdiff(t1, t2))
+       a   b
+      --- ---
+        4   d
+
+      >>> print(setdiff(t2, t1))
+       a   b
+      --- ---
+        5   b
+    """
+    if keys is None:
+        keys = table1.colnames
+
+    #Check that all keys are in table1 and table2
+    for tbl, tbl_str in ((table1,'table1'), (table2,'table2')):
+        diff_keys = np.setdiff1d(keys, tbl.colnames)
+        if len(diff_keys) != 0:
+            raise ValueError("The {} columns are missing from {}, cannot take "
+                             "a set difference.".format(diff_keys, tbl_str))
+
+    # Make a light internal copy of both tables
+    t1 = table1.copy(copy_data=False)
+    t1.meta = {}
+    t1.keep_columns(keys)
+    t1['__index1__'] = np.arange(len(table1))  # Keep track of rows indices
+
+    # Make a light internal copy to avoid touching table2
+    t2 = table2.copy(copy_data=False)
+    t2.meta = {}
+    t2.keep_columns(keys)
+    # Dummy column to recover rows after join
+    t2['__index2__'] = np.zeros(len(t2), dtype=np.uint8)  # dummy column
+
+    t12 = _join(t1, t2, join_type='left', keys=keys,
+                metadata_conflicts='silent')
+
+    # If t12 is masked then that means some rows were in table1 but not table2.
+    if t12.masked:
+        # Define bool mask of table1 rows not in table2
+        diff = t12['__index2__'].mask
+        # Get the row indices of table1 for those rows
+        idx = t12['__index1__'][diff]
+        # Select corresponding table1 rows straight from table1 to ensure
+        # correct table and column types.
+        t12_diff = table1[idx]
+    else:
+        t12_diff = table1[[]]
+
+    return t12_diff
 
 
 def vstack(tables, join_type='outer', metadata_conflicts='warn'):