+++ CHANGES.rst
@@ -79,6 +79,9 @@ astropy.table
 
 - Allow to pass ``htmldict`` option to the jsviewer writer. [#6551]
 
+- Added new table operation ``astropy.table.setdiff`` that returns the set
+  difference of table rows for two tables. [#6443]
+
 astropy.tests
 ^^^^^^^^^^^^^